Refactor : 디테일 태스크를 병렬 라우팅
