+++ src/app/(content-layout)/[groupId]/tasklist/@detailTask/_components/CommentField.tsx
@@ -4,8 +4,8 @@ import { Comment } from '@/components/comment/types';
 import { useState } from 'react';
 import EditCommentInput from './EditCommentInput';
 import axiosClient from '@/lib/axiosClient';
-import RemoveCommentModal from '../ModalContents/RemoveCommentModal';
 import useModalContext from '@/components/common/modal/core/useModalContext';
+import RemoveCommentModal from '../../_tasklist/components/ModalContents/RemoveCommentModal';
 
 interface Props {
   comment: Comment;
