+++ src/app/(content-layout)/[groupId]/tasklist/@detailTask/_components/DetailTaskContentField.tsx
@@ -1,12 +1,12 @@
 import DropDown from '@/components/common/dropdown';
-import { DetailTaskType } from '../../types/task-type';
 import Image from 'next/image';
 import ProfileBadge from '@/components/profile-badge';
 import Repeat from '@/assets/Repeat';
 import { format } from 'date-fns';
 import clsx from 'clsx';
-import { useTaskModals } from '../../hooks/use-task-modals';
-import { getRepeatDescription } from '../../utils/format-repeat-schedule';
+import { useTaskModals } from '../../_tasklist/hooks/use-task-modals';
+import { getRepeatDescription } from '../../_tasklist/utils/format-repeat-schedule';
+import { DetailTaskType } from '../../_tasklist/types/task-type';
 
 interface Props {
   task: DetailTaskType;
